requirements_sharepoint.txt
# Dépendances pour l'accès SharePoint et l'identification de fichiers DPGF/BPU/DQE
# Authentification et API Microsoft Graph
msal>=1.20.0
requests>=2.28.1

# Configuration et environnement
python-dotenv>=0.21.0

# Traitement des données et fichiers Excel
pandas>=1.5.2
openpyxl>=3.0.9
numpy>=1.21.0

# Base de données (si import en base requis)
sqlalchemy>=2.0.0

# Interface utilisateur et progress
tqdm>=4.62.3

# Validation des données
pydantic>=2.0.0

# IA pour l'analyse avancée (optionnel)
google-generativeai>=0.1.0

# Utilitaires système
pathlib2>=2.3.0  # Pour compatibilité Windows
urllib3>=1.26.0

# Logging et debugging
colorlog>=6.7.0  # Pour des logs colorés (optionnel)
